Shankari_SR_3122225001125_C/requirements.txt
# Core Dependencies
requests>=2.28.0
pymongo>=4.3.0
python-dotenv>=0.19.0

# Optional: For enhanced logging and monitoring
colorlog>=6.7.0

# Optional: For data validation
jsonschema>=4.17.0

# Optional: For enhanced date/time handling
python-dateutil>=2.8.0

# Development Dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0